Benchmarks: re-enable profiling-te configuration (try 2).
The previous PR was reverted since I didn't update the
`upload_scribe.py` script and it was looking for 'executor_and_fuser'
field in the json which now is replaced with two separate fields:
'executor' and 'fuser'.

[ghstack-poisoned]
