+++ benchmarks/upload_scribe.py
@@ -96,7 +96,8 @@ class PytorchBenchmarkUploader(ScribeUploader):
             test = b['name'].split('[')[0]
             net_name = b['params']['net_name']
             benchmark_name = '{}[{}]'.format(test, net_name)
-            executor, fuser = b['params']['executor_and_fuser'].split('-')
+            executor = b['params']['executor']
+            fuser = b['params']['fuser']
             m = self.format_message({
                 "time": upload_time,
                 "benchmark_group": b['group'],
